chore: Refactor some stuff
Group the api creation in apistack in its own method. Rename the
SNSTopic class, since it's actually the specific submissions topic. Move
that to its own class.
